refactor: [#15] ResultPage 비슷한 동물 skeleton 처리
