+++ src/pages/Test/TestPage.tsx
@@ -2,7 +2,7 @@ import { motion } from 'framer-motion';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import questions from '../data/questions.json';
+import questions from '../../data/questions.json';
 
 // type Question = {
 //   question: string;
